requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0  # For async HTTP in dispatcher
websockets>=12.0  # For WebSocket support in dashboard
